clamp min/max values by injecting into the simplex vector
Previously there was no feedback to the PDL::Opt::Simplex algorthim with
respect to min/max clamping.  We clamped the variable before passing it to f(),
but if simplex went way out of range it would just clamp and fail to iterate in
a useful way.

Now we `$vec .= <clamped vector>` to feed the clamped values back into the
calling simplex.  This way PDL::Opt::Simplex::Simple can feed the clamped
values back into the simplex algorithm so it will continue to try values near
the limit (instead of searching out of range).

This also fixes an initialization constraint bug: simplex breaks if an initial
value for a simplex variable is equal to the maximum constraint.  If the user
tries this, it will `die` with an error.