fix min/max clamping when var==max
PDL::Opt::Simplex uses a constant 0.00001 as (what appears to be) a minimum
step size.  If the variable being optimized is clamped to the upper limit on
the first iteration then it will fail almost immediately.  Thus, we clamp to
($var_max*1.0001) which seems to be just enough to prevent it from failing, and
probably isn't too much bigger than the requested constraint.