particle swarm: Add stallSearchScale
If a particle stalls, then jump to a random location that exists +/-
searchSize% of the current location, but increase searchSize% for that particle
by stallSearchScale after each stall.

This is based on a response to a question about how best to restart a particle
when it stalls:
	https://math.stackexchange.com/q/4753072/983059

  @BrianHopkins notes:

	"Both options seem to favor finding the same minima repeatedly. Based on a
	slightly earlier meta-heuristic, simulated annealing, you might restart a
	particle by having it jump some distance in a random direction from its local
	minimum (however you define direction and distance in the space). As time
	passes, the jump distances decrease."

This commit also fixes some debug print formatting and velocity calculation
debug prints.